jupyter/project/requirements.txt
Faker==30.3.0
faker-commerce==1.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
kafka-python==2.0.2
clickhouse-connect==0.8.3
requests==2.27.1
clickhouse-connect==0.8.3